Add Get-PuppetAuthToken
This commit adds the Get-PuppetAuthToken cmdlet.

It also adds some private functions that help with checking to see if
the -SkipCertificateCheck flag is supported in the Invoke-RestMethod
cmdlet in the current session, and if not, setting the
ServerCertificateValidationCallback setting, so that PS 5.1 can
connect to Puppet Masters that have an untrusted SSL cert.

The Set-ServerCertificateValidationCallback cmdlet is there purely as
a stub for mocking to make it easier to pester test that code path.